pyproject.toml
[project]
name = "webaiback"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic==1.18.3",
    "annotated-doc==0.0.4",
    "annotated-types==0.7.0",
    "anyio==4.12.1",
    "asyncpg==0.31.0",
    "beautifulsoup4==4.14.3",
    "certifi==2026.1.4",
    "cffi==2.0.0",
    "charset-normalizer==3.4.4",
    "click==8.3.1",
    "colorama==0.4.6",
    "cryptography==46.0.4",
    "distro==1.9.0",
    "dnspython==2.8.0",
    "dotenv==0.9.9",
    "ecdsa==0.19.1",
    "email-validator==2.3.0",
    "fastapi==0.128.0",
    "google-auth==2.48.0",
    "google-genai==1.60.0",
    "greenlet==3.3.1",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "idna==3.11",
    "langsmith==0.7.1",
    "librt==0.7.8",
    "mako==1.3.10",
    "markupsafe==3.0.3",
    "mypy==1.19.1",
    "mypy-extensions==1.1.0",
    "numpy==2.4.1",
    "orjson==3.11.7",
    "packaging==26.0",
    "passlib==1.7.4",
    "pathspec==1.0.4",
    "pgvector==0.4.2",
    "psycopg2-binary==2.9.11",
    "pyasn1==0.6.2",
    "pyasn1-modules==0.4.2",
    "pycparser==3.0",
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",
    "pydantic-settings==2.12.0",
    "pypdf==6.6.2",
    "python-dotenv==1.2.1",
    "python-jose==3.5.0",
    "python-multipart==0.0.22",
    "requests==2.32.5",
    "requests-toolbelt==1.0.0",
    "rsa==4.9.1",
    "six==1.17.0",
    "sniffio==1.3.1",
    "soupsieve==2.8.3",
    "sqlalchemy==2.0.46",
    "starlette==0.50.0",
    "tenacity==9.1.2",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "urllib3==2.6.3",
    "uuid-utils==0.14.0",
    "uvicorn==0.40.0",
    "websockets==15.0.1",
    "xxhash==3.6.0",
    "zstandard==0.25.0",
]
[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "B"]
ignore = [
    "E501", # Вимикаємо помилку занадто довгих рядків
    "B008", # Дозволяємо Depends() у параметрах FastAPI
]


[tool.ruff.lint.isort]
known-first-party = ["app"]


[tool.ruff.lint.extend-per-file-ignores]
"app/models/*" = ["N807"] # Дозволяємо __repr__ у моделях
"alembic/*" = ["F403"]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
]
